requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
langchain
faiss-cpu
gtts
speechrecognition
googletrans==4.0.0-rc1
pillow
streamlit-webrtc
